+++ triton_kernels/rmsnorm/rmsnorm_with_loops.py
@@ -28,4 +28,4 @@ def rmsnorm_kernel_1(output_ptr,input_ptr,w_ptr,eps,input_row_stride,BLOCK_SIZE:
       mask = cols < input_row_stride
       x_block = tl.load(X_row + cols, mask = mask, other = 0.0).to(tl.float32)
       w_block = tl.load(w_ptr + cols, mask=mask,other = 0.0).to(tl.float32)
-      tl.store(Y_row + cols, x_block * w_block * normfactor, mask=mask)
+      tl.store(Y_row + cols, (x_block * w_block * normfactor).to(tl.float16), mask=mask)
